Fix: compat07::SnapshotMeta should decode v08 SnapshotMeta
