+++ rocksstore-compat07/Cargo.toml
@@ -15,7 +15,7 @@ repository    = { workspace = true }
 
 
 [dependencies]
-openraft = { path = "../openraft", package = "openraft", version = "0.8.4", features = ["compat-07"] }
+openraft = { path = "../openraft", package = "openraft", version = "0.8.4", features = ["compat-07", "bt"] }
 
 async-std = { version = "1.12.0", features = ["attributes", "tokio1"] }
 byteorder = "1.4.3"
